Overhaul API
This is a major fundamental change to the API. Memcache.* commands no longer
take a connection pid/name or command batch as the first argument. The
expectation is that you're running a Memcache.Router. This makes the easy path
very straightforward. It also eliminates these functions having different modes
depending on how they're called.

To build up a batch of commands, look for equivalent functions under
Memcache.Command. These just return a command tuple. Put them in a list and
pass the list to Memcache.execute to get back a list of results.

The functions under StateMachine had a similar overhaul. They just return a
single machine tuple. Pass a list of these to StateMachine.run to get a list of
results.